Refactor utilities into their own namespace and split up utils class
