+++ PSLDiscordBot.Core/ImageGenerating/ImageGenerator.cs
@@ -1,5 +1,6 @@
 ﻿using HtmlToImage.NET;
 using Newtonsoft.Json;
+using PSLDiscordBot.Framework.Utilities;
 using SixLabors.ImageSharp;
 using SixLabors.ImageSharp.Formats.Png;
 using SixLabors.ImageSharp.PixelFormats;
