+++ PSLDiscordBot.Framework/BuiltInServices/CommandResolveService.cs
@@ -2,6 +2,7 @@
 using Discord.WebSocket;
 using PSLDiscordBot.Framework.CommandBase;
 using PSLDiscordBot.Framework.MiscEventArgs;
+using PSLDiscordBot.Framework.Utilities;
 using System.Collections.Concurrent;
 using System.Reflection;
