+++ PSLDiscordBot.Framework/Localization/LocalizationManager.cs
@@ -1,4 +1,5 @@
 ﻿using Newtonsoft.Json;
+using PSLDiscordBot.Framework.Utilities;
 
 namespace PSLDiscordBot.Framework.Localization;
